requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==4.0.1
unstructured==0.21.5
pdf2image==1.16.3
pillow==10.1.0
pdfplumber==0.10.4
chromadb==0.5.0
sentence-transformers==3.0.1
torch==2.1.2
requests==2.31.0